Fix removal of i.e.
The regex for `i.e.` currently isn't checking for the period character,
it's checking for `i<any character>e<any character>`. This PR adds an
example to the "Random addresses that have broken this library at some
point." section of tests and updates the regex to look for period
characters instead of any character.